Moved things around.
I moved a lot of things around to incorporate the boa packaging and to
make the server make a bit more sense. Expect the same to be done to the
Central Server and even more moving done to this server as well as
deleting a lot of "content" as I strip this server down to a core to
rebuild it.