Updated the Connect class to use my logger.
I updated all println and logger references in Connect to use my logger.
This is probably a huge mistake and I'll probably revert back to the
standard library logger. My logger's only advantages are that I've made
it add an extra line so that prints are spaced out and you only need the
one instance of the class rather than an instance in every class you
want a proper print in such as with the standard library logger.

Also BIG TODO: I really really really want to fix that Connect class.
It's disgusting and really needs to be at least organized, if not ripped
out entirely and replaced with a netty alternative.