Changed Paul's packet system
I fixed up Paul's old packet system that he made during the RS-Reborn
days. It was a great system that, for once, appeared to follow
RuneScape's trend (as most things when it comes to Paul don't). It was
easy to implement AS HANDLERS, but it was extremely difficult to
implement new packets into the system itself. The entire ordeal
consisted of creating an interface for the method, implementing 4
basically identical methods into each of the four priority classes and
then adding another method to the main class, Content, which is what you
actually called the methods from. I took another look at it (when I knew
what I was doing) and noticed that, while I didn't see any memory LEAKS,
there was a lot of, frankly unnecessary, LinkedLists strewn about. There
was one for EACH interface in ALL FOUR of the handler priorities! This
means if you had 20 interfaces, you have to multiply that number by four
to get the number of LinkedLists, which ends up being 80. Oy vey.

So in my new version of the handler system, the methods are a little
more complicated to add, but there's only one each (and they aren't THAT
complicated, and I may make them easier in the future) and while you
still have to make the interface, I've kept the entire system
implementation in handlers exactly the same.